requirements.txt
google-cloud-bigtable
google-cloud-core
